requirements.txt
ruff==0.1.1
aiogram==3.1.1
python-dotenv==1.0.0
StrEnum==0.4.15
# psycopg2==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
